refactor(core-ui): move rememberAnimationsEnabled to core.ui.motion (review)
Codex gate on #754: SkeletonBox (core.ui.loading) depended on
core.ui.post for the reduce-motion helper — an inverted package
dependency. The helper is a shared motion concern (post shimmers, pull
loader, loading skeletons), now under core.ui.motion. Also a plain Dp
import in TopicLoadingSkeleton instead of an inline FQN.
