+++ core/ui/src/main/kotlin/fr/forumhfr/redface2/core/ui/post/ImageShimmer.kt
@@ -1,10 +1,7 @@
 package fr.forumhfr.redface2.core.ui.post
 
-import android.provider.Settings
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import fr.forumhfr.redface2.core.ui.loading.SkeletonBox
 
 /**
@@ -22,17 +19,3 @@ internal fun ImageShimmer(animated: Boolean, modifier: Modifier = Modifier) {
     SkeletonBox(modifier = modifier, animated = animated)
 }
 
-/**
- * #249 §4 — `true` unless the user disabled animations system-wide (Developer options / "Remove
- * animations", or an accessibility profile setting `ANIMATOR_DURATION_SCALE` to 0). Reading the system
- * setting (rather than the app's unwired `future_a11y_reduce_motion` row) honours the OS-level
- * preference the same way Compose's own animations do. Defaults to enabled if the setting is absent.
- */
-@Composable
-internal fun rememberAnimationsEnabled(): Boolean {
-    val resolver = LocalContext.current.contentResolver
-    return remember(resolver) {
-        val scale = Settings.Global.getFloat(resolver, Settings.Global.ANIMATOR_DURATION_SCALE, 1f)
-        scale != 0f
-    }
-}
